Return early from key handler if key typed is not of interest
This commit removes a clause which ignored keyboard events when a
modifier key was pressed (ctrl, alt, meta) and replaces it with a
clause which returns early if the key pressed is not of interest
to keyzen.

The previous behaviour prevented typing the '#' on UK Mac keyboards
which is typed with alt-3.

This fixes #7.